requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
gitpython>=3.1.0
requests>=2.31.0
